Remove cruft, move better version of this into run.py.
